chore/Refactor get_aad_users method
